// src/modules/utils/enums.ts
export enum Chain {
  ELA = 'ela',
  ETH = 'eth',
  FSN = 'fsn',
  V1 = 'v1',
}

export enum Category {
  General = 'general',
  Art = 'art',
  Collectibles = 'collectibles',
  Photography = 'photography',
  TradingCards = 'trading cards',
  Utility = 'utility',
  Domain = 'domain',
}

export enum OrderTag {
  BuyNow = 0,
  OnAuction = 1,
  NotMet = 2,
  HasBids = 3,
  HasEnded = 4,
}